+++ tools/gmail/src/gmail.ts
@@ -51,7 +51,7 @@ import {
  *     const authLink = await this.gmail.requestAuth(this.onGmailAuth);
  *
  *     await this.plot.createActivity({
- *       type: ActivityType.Task,
+ *       type: ActivityType.Action,
  *       title: "Connect Gmail",
  *       links: [authLink]
  *     });